requirements.txt
flask
pandas
python-nmap
requests
gunicorn

